nexus: initial integration of new gp code with nexus
