+++ nexus/lib/gaussian_process.py
@@ -23,35 +23,86 @@
 #      Function to calculate the pair-wise point distances of a      #
 #      matrix.  Subroutine of lhsmaximin.                            #
 #                                                                    #
+#    _fdist                                                          #
+#      Function to calculate distances between all points.           #
+#      Subroutine of find_hyperparameters, find_min, and             #
+#      find_new_points.                                              #
+#                                                                    #
 #    cfgp                                                            #
 #      Function to calculate correlation function for Gaussian       #
 #      Process.  Subroutine of cmgp.                                 #
 #                                                                    #
 #    cmgp                                                            #
 #      Function to calculate correlation matrix for Gaussian         #
-#      Process.  Subroutine of sgd and gp_sgd.                       #
+#      Process.  Subroutine of find_min.                             #
 #                                                                    #
-#    sgd                                                             #
-#      Function to calculate minimums of Gaussian Process using      #
-#      stochastic gradient descent.  Subroutine of gp_sgd.           #
+#    find_hyperparameters                                            #
+#      Function to find best Gaussian Process hyperparameters.       #
+#      Subroutine of gp_opt.                                         #
 #                                                                    #
-#    gp_sgd                                                          #
+#    find_min                                                        #
+#      Function to find local minimum of Gaussian Process surface.   #
+#      Subroutine of gp_opt.                                         #
+#                                                                    #
+#    find_new_points                                                 #
+#      Function to estimate optimal new points for Gaussian Process. #
+#      Subroutine of gp_opt.                                         #
+#                                                                    #
+#    multiscale_E                                                    #
+#      Function to perform multiscale normalization of E.            #
+#      Subroutine of gp_opt.                                         #
+#                                                                    #
+#    inv_multiscale_E                                                #
+#      Function to reverse multiscale normalization of E.            #
+#      Subroutine of gp_opt.                                         #
+#                                                                    #
+#    norm_pos                                                        #
+#      Function to normalize molecular positions to unit hyper cube. #
+#      Subroutine of gp_opt.                                         #
+#                                                                    #
+#    inv_norm_pos                                                    #
+#      Function to reverse normalization of molecular positions.     #
+#      Subroutine of gp_opt.                                         #
+#                                                                    #
+#    gp_opt_setup                                                    #
+#      Function that sets up parameters and variables for GP.        #
+#                                                                    #
+#    gp_opt                                                          #
 #      Function that performs one iteration of Gaussian Process      #
-#      optimization using stochastic gradient descent.  This is the  #
-#      main interface to Gaussian Process iterations.                #
+#      optimization.  This is the main interface to Gaussian Process #
+#      iterations.                                                   #
 #                                                                    #
-#    ackleyf                                                         #
-#      Function to calculate the d-dimensional Ackley function on    #
-#      [-1,1]^d.                                                     #
+#    opt_surface_f                                                   #
+#      Example function to optimize.                                 #
 #                                                                    #
-#    ackley_example                                                  #
-#      Example using Gaussian Process optimization on Ackley         #
-#      function.                                                     #
+#    gp_opt_example                                                  #
+#      Example using Gaussian Process optimization on opt_surface_f. #
 #                                                                    #
 #====================================================================#
 
 
+from developer import unavailable,available,DevBase,log,warn,error
+
 import numpy as np
+try:
+    import matplotlib.pyplot as plt
+except:
+    plt = unavailable('matplotlib.pyplot')
+#end try
+try:
+    from scipy.spatial import Delaunay
+except:
+    Delaunay = unavailable('scipy.spatial','Delaunay')
+#end try
+
+
+class GParmFlags(DevBase):
+    """
+    Define an empty class to store GP parameters and information
+    """
+    pass
+#end class GParmFlags
+
 
 
 def lhs(n, samples=None, iterations=None):
@@ -116,14 +167,12 @@ def _lhsclassic(n, samples):
     for j in range(n):
         rdpoints[:, j] = u[:, j]*(b-a) + a
     #end for
-
     # Make the random pairings
     H = np.zeros_like(rdpoints)
     for j in range(n):
         order = np.random.permutation(range(samples))
         H[:, j] = rdpoints[order, j]
     #end for
-
     return H
 #end def _lhsclassic
     
@@ -141,7 +190,6 @@ def _lhsmaximin(n, samples, iterations):
             maxdist = np.min(d)
             H = Hcandidate.copy()
         #end if
-    #end for
     return H
 #end def _lhsmaximin
 
@@ -185,7 +233,6 @@ def _pdist(x):
     if m<2:
         return []
     #end if
-    
     d = np.zeros((m**2-m)/2)
     c_v = 0
     for i in range(m - 1):
@@ -193,37 +240,59 @@ def _pdist(x):
                         [1,n]),m-i-1,axis=0)-x[i+1:m+1, :])**2,axis=1))
         c_v = c_v+m-i-1
     #end for
-
     return d
 #end def _pdist
 
 ###############################################################################
     
-def ackleyf(x):
+def _fdist(x):
     """
-    Calculate the d-dimensional Ackley function on [-1,1]^d
+    Calculate Distance between all points
     
-    see: Floudas, C. A., & Pardalos, P. M. (1990). A collection of test 
-    problems for constrained global optimization algorithms. In G. Goods & 
-    J. Hartmanis (Eds.) Lecture notes in computer science: Vol. 455. 
-    Berlin: Springer
     
+    Parameters
+    ----------
+    x: n x d array of points
+
+    Returns
+    -------
+    Dm : n x m correlation matrix
+    
+    """
+    
+    [n,d] = x.shape
+    Dm = np.zeros([n,n])
+    
+    for jn in range(n):
+        Dm[jn,:] = np.sqrt(np.sum((np.repeat(np.reshape(x[jn,:],[1,d]),\
+             n,axis=0)-x)**2,axis=1))
+    #end for
+    
+    return Dm
+#end def _fdist
+
+###############################################################################
+    
+def opt_surface_f(x):
+
+    """
+    Optimization Surface Function 
+        
     Parameters
     ----------
     x : 1 x d-array, x \in [-1,1]^d
     
     Returns
     -------
-    y : value of Ackely function
+    y : value of optimization surface function
     
     """
-    d = x.shape[1]
-    x = 0.5*x
-    y = 20.0+np.exp(1.0)-20.0*np.exp(-0.2*np.sqrt(np.sum(x**2)/float(d)))- \
-        np.exp(np.sum(np.cos(2.0*np.pi*x))/float(d))
-    
+    [n,d] = x.shape
+    y = np.reshape(2.0*np.sin(np.pi*np.sum(x**2,1)/2)-1,[n,1])
+    #y = np.reshape(np.prod(np.sin(2*np.pi*x**2),axis=1),[n,1])
+
     return y
-#end def ackleyf
+#end def opt_surface
 
 ###############################################################################
 
@@ -276,7 +345,6 @@ def cmgp(S1,S2,ep):
         K[jn,:] = np.sqrt(np.sum((np.repeat(np.reshape(S1[jn,:],[1,d]),\
              m,axis=0)-S2)**2,axis=1))
     #end for
-    
     K = cfgp(K,ep)
     
     return K
@@ -284,302 +352,675 @@ def cmgp(S1,S2,ep):
 
 ###############################################################################
 
-def sgd(hyper_cube,co_GP,mol_pos,ep,tol=None,verbose=None):
+def find_hyperparameters(mol_pos,E,dE):
     """
-    Calculate minimums of Gaussian Process using stochastic gradient descent
+    Finds best Gaussian Process hyperparameters
     
     
     Parameters
     ----------
-    co_GP:      GP coefficients
     mol_pos:    GP centers 
-    ep:         GP kernel parameter ep>0
-    tol:        Tolerance for termination
+    E:          Energy for GP centers
+    dE:         Uncertainty in energy
     
     Returns
     -------
-    min_loc : Local minimums
-    min_E   : Energy Estimation of local minimums
+    ep,sigma:   GP hyperparameters
+    Co_GP:      GP coefficents
     """
+    [Nc,d] = mol_pos.shape
     
-    if tol is None:
-        tol = 1e-4
+    W = np.diag(dE[:,0])
+    
+    D =_fdist(mol_pos)
+    ep = .5/np.max(np.min(D+np.max(D)*np.eye(Nc),axis=1))**2
+    ep_ceil = 10*ep
+    d_ep = 1e-2*ep
+    tol = 1e-3*ep
+    
+    sigma = 1.0
+    d_sigma = 1e-2*sigma
+    tol_sigma = 1e-3*sigma
+    ep_c = ep
+    
+    K = cfgp(D,ep_c)
+    Kw_inv = np.linalg.pinv(K+ sigma*W)
+    Co_GP = np.matmul(Kw_inv,E)
+    loocv = Co_GP/np.reshape(np.diag(Kw_inv),[Nc,1])
+    
+    if sigma == 0:
+        b_loocv_err = np.sum(-(loocv**2))
+        d_sigma = 0
+    else:
+        sigma_calc = 1-np.diag(np.matmul(K,np.matmul(Kw_inv,np.transpose(K))))
+        sigma_calc = np.reshape(np.maximum(sigma_calc,(sigma/10)**2),[Nc,1])
+        b_loocv_err = np.sum(-(loocv**2)/sigma_calc-np.log(sigma_calc)/2)- \
+            Nc*np.sum(np.abs(np.abs(E-np.matmul(K,Co_GP))-dE)/dE)
     #end if
-    
-    if verbose is None:
-        verbose=0
-    #end if
+            
+    while d_ep>tol and d_sigma>tol_sigma:
+        did_move = 0;
+        ep_c = min(ep+d_ep,ep_ceil)
+        K = cfgp(D,ep_c)
+        Kw_inv = np.linalg.pinv(K+ sigma*W)
+        Co_GP = np.matmul(Kw_inv,E)
+        loocv = Co_GP/np.reshape(np.diag(Kw_inv),[Nc,1])
+        if sigma == 0:
+            sigma_calc = 1;
+            c_loocv_err = np.sum(-(loocv**2))
+        else:
+            sigma_calc = 1-np.diag(np.matmul(K,np.matmul(Kw_inv,np.transpose(K))))
+            sigma_calc = np.reshape(np.maximum(sigma_calc,(sigma/10)**2),[Nc,1])
+            c_loocv_err = np.sum(-(loocv**2)/sigma_calc-np.log(sigma_calc)/2)- \
+                Nc*np.sum(np.abs(np.abs(E-np.matmul(K,Co_GP))-dE)/dE)
+        #end if 
+            
+        if c_loocv_err>b_loocv_err:
+            b_loocv_err = c_loocv_err
+            ep = ep_c
+            did_move = 1
+        #end if
 
-    npoints = mol_pos.shape[0]
-    d = mol_pos.shape[1]
-    epm = np.min(_pdist(mol_pos))
-    
-    best_pos = np.copy(mol_pos)
-    delta_pos = epm*np.ones([npoints,d])/10.0
-    best_E = np.matmul(cmgp(best_pos,mol_pos,ep),co_GP)
-    
-    cm = np.max(delta_pos)
-    use_ind = np.where(np.max(delta_pos,axis=1)>epm*tol)
-    while cm>epm*tol:
-        for jd in range(d):
-            did_move = np.zeros([npoints,1])
-            t_pos =  np.copy(best_pos)
-            t_pos[:,jd] = t_pos[:,jd] + delta_pos[:,jd]
-            for jdim in range(d):
-                t_pos[np.where(t_pos[:,jdim]>hyper_cube[jdim,1]),jdim]= \
-                    hyper_cube[jdim,0]
-            #end for
+        ep_c = ep-d_ep
+        if ep_c<=0:
+            ep_c = ep
+        #end if
 
-            c_E = np.matmul(cmgp(t_pos,mol_pos,ep),co_GP)
-            for jp in range(npoints):
-                if c_E[jp,0]<best_E[jp,0]:
-                    did_move[jp,0] = 1
-                    best_E[jp,0] = c_E[jp,0]
-                    best_pos[jp,:] = t_pos[jp,:]
+        K = cfgp(D,ep_c)
+        Kw_inv = np.linalg.pinv(K+ sigma*W)
+        Co_GP = np.matmul(Kw_inv,E)
+        loocv = Co_GP/np.reshape(np.diag(Kw_inv),[Nc,1])
+        if sigma == 0:
+            sigma_calc = 1;
+            c_loocv_err = np.sum(-(loocv**2))
+        else:
+            sigma_calc = 1-np.diag(np.matmul(K,np.matmul(Kw_inv,np.transpose(K))))
+            sigma_calc = np.reshape(np.maximum(sigma_calc,(sigma/10)**2),[Nc,1])
+            c_loocv_err = np.sum(-(loocv**2)/sigma_calc-np.log(sigma_calc)/2)- \
+                Nc*np.sum(np.abs(np.abs(E-np.matmul(K,Co_GP))-dE)/dE)
+        #end if
+            
+        if c_loocv_err>b_loocv_err:
+            b_loocv_err = c_loocv_err
+            ep = ep_c
+            did_move = 1
+        #end if
+
+        if did_move == 1:
+            d_ep = 1.25*d_ep
+        else:
+            d_ep = .75*d_ep
+        #end if
+
+        if sigma>0:
+            did_move = 0;
+            sigma_c = sigma+d_sigma
+            K = cfgp(D,ep)
+            Kw_inv = np.linalg.pinv(K+ sigma_c*W)
+            Co_GP = np.matmul(Kw_inv,E)
+            loocv = Co_GP/np.reshape(np.diag(Kw_inv),[Nc,1])
+
+            sigma_calc = 1-np.diag(np.matmul(K,np.matmul(Kw_inv,np.transpose(K))))
+            sigma_calc = np.reshape(np.maximum(sigma_calc,(sigma/10)**2),[Nc,1])
+            c_loocv_err = np.sum(-(loocv**2)/sigma_calc-np.log(sigma_calc)/2)- \
+                Nc*np.sum(np.abs(np.abs(E-np.matmul(K,Co_GP))-dE)/dE)
+                
+                
+            if c_loocv_err>b_loocv_err:
+                b_loocv_err = c_loocv_err
+                sigma = sigma_c
+                did_move = 1
+            #end if
+ 
+            sigma_c = sigma-d_sigma
+            if sigma_c<=0:
+                sigma_c = sigma
+            #end if
+
+            Kw_inv = np.linalg.pinv(K+ sigma_c*W)
+            Co_GP = np.matmul(Kw_inv,E)
+            loocv = Co_GP/np.reshape(np.diag(Kw_inv),[Nc,1])
+
+            sigma_calc = 1-np.diag(np.matmul(K,np.matmul(Kw_inv,np.transpose(K))))
+            sigma_calc = np.reshape(np.maximum(sigma_calc,(sigma/10)**2),[Nc,1])
+            c_loocv_err = np.sum(-(loocv**2)/sigma_calc-np.log(sigma_calc)/2)- \
+                Nc*np.sum(np.abs(np.abs(E-np.matmul(K,Co_GP))-dE)/dE)
+                
+                
+            if c_loocv_err>b_loocv_err:
+                b_loocv_err = c_loocv_err
+                sigma = sigma_c
+                did_move = 1
+            #end if
+
+            if did_move == 1:
+                d_sigma = 1.25*d_sigma
+            else:
+                d_sigma = .75*d_sigma
+            #end if
+        #end if
+        K = cfgp(D,ep)
+        Kw_inv = np.linalg.pinv(K+ sigma*W)
+        Co_GP = np.matmul(Kw_inv,E)
+    #end while
+
+    return ep,sigma,Co_GP 
+#end def find_hyperparameters
+
+###############################################################################
+
+def find_min(Co_GP,mol_pos,E,dE,ep):
+    """
+    Finds local minimum of Gaussian Process
+    
+    
+    Parameters
+    ----------
+    mol_pos:    GP centers 
+    E:          Energy for GP centers
+    dE:         Uncertainty in energy for GP centers
+    Co_GP:      GP coefficents
+    ep:         GP exponential parameter
+    
+    Returns
+    -------
+    best_pos:   Estimated local minimum
+    best_E:     Estimated local energies
+    neig_data:  Neighborhood Data
+    """
+    [Nc,d] = mol_pos.shape
+    
+    T = Delaunay(mol_pos)
+    D =_fdist(mol_pos)
+    best_pos = mol_pos
+    ref_dis = np.min(np.min(D+np.max(D)*np.eye(Nc),axis=1))/10
+    K = cfgp(D,ep)
+    
+    delta_pos = ref_dis+np.zeros([Nc,d])
+    tol_pos = ref_dis/10
+    best_E = np.matmul(K,Co_GP)
+    
+    while np.max(delta_pos)>tol_pos:
+        for jp in range(Nc):
+            for jd in range(d):
+                did_move = 0
+                c_pos = np.copy(best_pos)
+                c_pos[jp,jd]= best_pos[jp,jd]+delta_pos[jp,jd]
+                ind_T = T.find_simplex(c_pos[jp,:])
+                if ind_T == -1:
+                    c_pos[jp,jd]= best_pos[jp,jd]
                 #end if
-            #end for
 
-            t_pos =  np.copy(best_pos)
-            t_pos[:,jd] = t_pos[:,jd] - delta_pos[:,jd]
-            for jdim in range(d):
-                t_pos[np.where(t_pos[:,jdim]<hyper_cube[jdim,0]),jdim]= \
-                    hyper_cube[jdim,0]
-            #end for
-
-            c_E = np.matmul(cmgp(t_pos,mol_pos,ep),co_GP)
-            for jp in range(npoints):
-                if c_E[jp,0]<best_E[jp,0]:
-                    did_move[jp,0] = 1
-                    best_E[jp,0] = c_E[jp,0]
-                    best_pos[jp,:] = t_pos[jp,:]
+                Kc = cmgp(np.reshape(c_pos[jp,:],[1,d]),mol_pos,ep)
+                c_E = np.matmul(Kc,Co_GP)
+                if c_E<best_E[jp]:
+                    best_E[jp] = c_E
+                    best_pos = np.copy(c_pos)
+                    did_move = 1
                 #end if
-                if did_move[jp,0] == 1:
+
+                c_pos = np.copy(best_pos)
+                c_pos[jp,jd]=best_pos[jp,jd]-delta_pos[jp,jd]
+                ind_T = T.find_simplex(c_pos[jp,:])
+                if ind_T == -1:
+                    c_pos[jp,jd]= best_pos[jp,jd]
+                #end if
+                Kc = cmgp(np.reshape(c_pos[jp,:],[1,d]),mol_pos,ep)
+                c_E = np.matmul(Kc,Co_GP)
+                if c_E<best_E[jp]:
+                    best_E[jp] = c_E
+                    best_pos = np.copy(c_pos)
+                    did_move = 1
+                #end if
+
+                if did_move == 1:
                     delta_pos[jp,jd] = 1.25*delta_pos[jp,jd]
                 else:
                     delta_pos[jp,jd] = 0.75*delta_pos[jp,jd]
                 #end if
             #end for
         #end for
-
-        cm = np.max(delta_pos)
-        if verbose == 1:
-            use_ind = np.where(np.max(delta_pos,axis=1)>epm*tol)
-            print cm,np.shape(use_ind)[1]
+    #end while
+     
+    ## Remove in same simplex ##
+    jp = 0
+    ind_T = T.find_simplex(best_pos)
+    ind_T = np.reshape(ind_T,[ind_T.shape[0],1])
+    while jp<best_pos.shape[0]:
+        us_ind = np.argwhere(ind_T==ind_T[jp])
+        if us_ind.shape[0]>1:
+            min_ind = np.argmin(best_E[us_ind[:,0],0])
+            best_E[us_ind[0,0],0] = best_E[us_ind[min_ind,0],0]
+            best_pos[us_ind[0,0],:] = best_pos[us_ind[min_ind,0],:]
+            best_E = np.delete(best_E,us_ind[1:us_ind.shape[0],0],0)
+            best_pos = np.delete(best_pos,us_ind[1:us_ind.shape[0],0],0)
+            ind_T = np.delete(ind_T,us_ind[1:us_ind.shape[0],0],0)
         #end if
+        jp += 1
     #end while
-
-    c_class = 0
-    ind_marker = -np.ones([npoints,1])
-    while np.min(ind_marker)<0:
-        indc = np.where(ind_marker<0)
-        cdata = best_pos[indc[0],:]
-        cvec = np.reshape(cdata[0,:],[1,d])
-        m = cdata.shape[0]
-        dis_v = np.sqrt(np.sum((np.repeat(cvec,m,axis=0)-cdata)**2,axis=1))
-        class_ind = indc[0][np.where(dis_v<10*epm*tol)]
-        ind_marker[class_ind,:] = c_class
-        c_class = c_class+1
+    jp = 0
+    while jp<best_pos.shape[0]:
+        c_dis = np.sqrt(np.sum((np.repeat(np.reshape(best_pos[jp,:],[1,d]),\
+             best_pos.shape[0],axis=0)-best_pos)**2,axis=1))
+        us_ind = np.argwhere(c_dis<ref_dis)
+        if us_ind.shape[0]>1:
+            min_ind = np.argmin(best_E[us_ind[:,0],0])
+            best_E[us_ind[0,0],0] = best_E[us_ind[min_ind,0],0]
+            best_pos[us_ind[0,0],:] = best_pos[us_ind[min_ind,0],:]
+            best_E = np.delete(best_E,us_ind[1:us_ind.shape[0],0],0)
+            best_pos = np.delete(best_pos,us_ind[1:us_ind.shape[0],0],0)
+        #end if
+        jp += 1
     #end while
-
-    min_loc = np.zeros([c_class,d])
-    min_E = np.zeros([c_class,1])  
+    ind_T = T.find_simplex(best_pos)
+    ind_T = np.reshape(ind_T,[ind_T.shape[0],1])
     
-    for jc in range(c_class):
-        indc = np.where(ind_marker==jc)
-        c_E = best_E[indc[0],:]
-        t_pos = best_pos[indc[0],:]
-        indb = np.argmin(c_E)
-        min_E[jc,:] = c_E[indb,:]
-        min_loc[jc,:] = t_pos[indb,:]
-    #end for
+    if best_E.shape[0] > 1:
+        ind_sort = np.argsort(best_E[:,0])
+        ind_rm = np.argwhere(best_E[ind_sort,0]>np.mean(E)-np.std(E))
+        if ind_rm.shape[0]>0:
+            if ind_rm[0,0] == 0:
+                ind_rm = 1
+            else:
+                ind_rm = ind_rm[0,0]
+            #end if  
+            ind_sort = ind_sort[0:ind_rm]
+        #end if
+        best_E = best_E[ind_sort,:]
+        best_pos = best_pos[ind_sort,:]
+        ind_T = ind_T[ind_sort,:]
+    #end if
+    neig_data = E[T.simplices[ind_T[:,0],:],0]
+    neig_data = np.append(neig_data,dE[T.simplices[ind_T[:,0],:],0],axis=0)
         
-    return min_loc,min_E
-#end def sgd
+    return best_pos,best_E,neig_data
+#end def find_min
 
 ###############################################################################
-def gp_sgd(hyper_cube_F,E_F,P_F,E0_F,P0_F,sigma,npoints=None,nlhc_its=None,\
-           ep=None,tol=None,verbose=None,gen_new=None):
+
+def find_new_points(mol_pos,best_pos,delta_r,Nc0):
     """
-    Performs one iteration of Gaussian Process optimization using 
-    stochastic gradient descent
+    Estimate optimal new points for Gaussian Process
     
     
     Parameters
     ----------
+    best_pos:   Estimated local minimum
+    best_E:     Estimated local energies
+    delta_r:    Maximum distance of new search space
+    Nc0:        Number of points to add
     
-    hyper_cube_F:   Hypercube for all levels
-    E_F:            Energies for all levels
-    P_F:            Sample points for all levels
-    E0_F:           Estimated Minimum Energy
-    P0_F:           Estimated Position of Minimum
-    sigma:          Error measure
-    npoints:        Number of points to create optimization surface 
-    nlhc_its:       The number of iterations in the lhc algorithm
-    ep:             GP kernel parameter ep>0
-    tol:            Tolerance for termination
-    
-    Returns following for next iteration
+    Returns
     -------
-    hyper_cube_F:   Hypercube for all levels
-    E_F:            Energies for all levels
-    P_F:            Sample points for all levels
-    E0_F:           Estimated Minimum Energy
-    P0_F:           Estimated Position of Minimum
+    new_points: New sample points for GP
     """
     
-    d = hyper_cube_F[0].shape[0]
+    [Nc,d] = mol_pos.shape
+    new_points = np.zeros([Nc0,d])
+    n_clusters = best_pos.shape[0]
+    ind_clusters = np.round(np.linspace(0,n_clusters,n_clusters+1)*Nc0/(n_clusters))
+    ind_clusters = ind_clusters.astype(int)  
+    shyper_cube = np.zeros([d,2])
+    n_min_d = 5
     
-    if tol is None:
-        tol = 1e-4
-    #end if
-    if verbose is None:
-        verbose=0 
-    #end if
-    if npoints is None:
-        npoints = (d+2)*(d+1)+1 
-    #end if
-    if nlhc_its is None:
-        nlhc_its = 10
-    #end if
-    if gen_new is None:
-        gen_new = 1
-    #end if
-
-    if len(P_F)==0:
-        hyper_cube = hyper_cube_F[0]
+    for jc in range(n_clusters):
+        n_pnts = ind_clusters[jc+1]-ind_clusters[jc]
+        shyper_cube[:,0] = np.maximum(best_pos[jc,:]-delta_r,0)
+        shyper_cube[:,1] = np.minimum(best_pos[jc,:]+delta_r,1)
+        test_points = np.repeat(np.reshape(shyper_cube[:,1]-shyper_cube[:,0],[1,d]) \
+            ,n_pnts,axis=0)*lhs(d,n_pnts)+ \
+            np.repeat(np.reshape(shyper_cube[:,0],[1,d]),n_pnts,axis=0)
+        test_points[0,:] = best_pos[jc,:]   
+        new_points[ind_clusters[jc]:ind_clusters[jc]+n_pnts,:] = test_points
         
-        ## Generate random sample from LHS
-        mol_pos = np.repeat(np.reshape(hyper_cube[:,1]-hyper_cube[:,0],[1,d]) \
-            ,npoints,axis=0)*lhs(d, samples=npoints,iterations=nlhc_its)+ \
-            np.repeat(np.reshape(hyper_cube[:,0],[1,d]),npoints,axis=0)
+        D =_fdist(np.append(mol_pos,new_points[0:ind_clusters[jc]+n_pnts,:],axis=0))
+        b_mes = np.sum(np.min(D+np.max(D)*np.eye(Nc+ind_clusters[jc]+n_pnts),axis=1))
         
-        hyper_cube[:,0] = np.reshape(np.min(mol_pos,axis=0),d)
-        hyper_cube[:,1] = np.reshape(np.max(mol_pos,axis=0),d)
-        
-        hyper_cube_F[0] = hyper_cube.copy()
-        
-        P_F.append(mol_pos)
-
-    else:
-        jits = len(P_F)-1
-        
-        mol_pos = P_F[jits]
-        E = E_F[jits]
-        hyper_cube = hyper_cube_F[jits]
-        ## Estimate Optimal Choice for ep
-        ep  = 1.0/np.max(_pdist(mol_pos))
-
-        ## Build Gaussian Process
-        K = cmgp(mol_pos,mol_pos,ep) 
-        Ks = K + (sigma**2)*np.identity(npoints)
-        co_GP = np.linalg.solve(Ks,E)
-        
-        if verbose != 0:
-            print "Starting Stochastic Grandient Descent"
-        #end if
-
-        ## Use stochastic gradient descent to determine minimums
-        min_loc,min_E = sgd(hyper_cube,co_GP,mol_pos,ep)
-        
-        indb = np.argmin(min_E)
-        bmin_loc = np.reshape(min_loc[indb,:],[1,d])
-        
-        P0_F.append(bmin_loc)
-        E0_F.append(min_E[indb])
-        
-        if gen_new == 1:
-            hyper_cube_F.append((hyper_cube_F[jits].copy()+\
-                        np.repeat(np.reshape(bmin_loc,[d,1]),2,axis=1))/2)
+        for jit in range(n_min_d):
+            cnew_points = np.copy(new_points)
+            test_points = np.repeat(np.reshape(shyper_cube[:,1]-shyper_cube[:,0],[1,d]) \
+                ,n_pnts,axis=0)*lhs(d,n_pnts)+ \
+                np.repeat(np.reshape(shyper_cube[:,0],[1,d]),n_pnts,axis=0)
+            test_points[0,:] = best_pos[jc,:]      
+            cnew_points[ind_clusters[jc]:ind_clusters[jc]+n_pnts,:] = test_points
             
-    
-            hyper_cube_F[jits+1][:,0] = np.maximum(hyper_cube_F[jits+1][:,0],\
-                        hyper_cube_F[jits][:,0])
-            hyper_cube_F[jits+1][:,1] = np.minimum(hyper_cube_F[jits+1][:,1],\
-                        hyper_cube_F[jits][:,1])
-            cLHC = hyper_cube_F[jits+1]
-            mol_pos = np.repeat(np.reshape(cLHC[:,1]-cLHC[:,0],[1,d])\
-                ,npoints,axis=0)*lhs(d, samples=npoints,iterations=nlhc_its)+\
-                np.repeat(np.reshape(cLHC[:,0],[1,d]),npoints,axis=0)
+            D =_fdist(np.append(mol_pos,cnew_points[0:ind_clusters[jc]+n_pnts,:],axis=0))
+            c_mes = np.sum(np.min(D+np.max(D)*np.eye(Nc+ind_clusters[jc]+n_pnts),axis=1))
             
-            P_F.append(mol_pos)
-            
-            hyper_cube_F[jits+1][:,0] = np.reshape(np.min(mol_pos,axis=0),d)
-            hyper_cube_F[jits+1][:,1] = np.reshape(np.max(mol_pos,axis=0),d)
-        #end if
-
-        if verbose != 0:
-            print "Error in min position",\
-                np.sqrt(np.sum((bmin_loc-x_true)**2)),"its=",jits
-                
-            if d == 1:
-                import matplotlib.pyplot as plt
-                npts = 100
-                mol_pos = P_F[jits]
-                plot_points = np.reshape(np.linspace(hyper_cube[0,0],\
-                    hyper_cube[0,1],num=npts),[npts,1])
-                gp_E = np.matmul(cmgp(plot_points,mol_pos,ep),co_GP)
-                Et = np.zeros([npts,1])
-                for jp in range(npts):
-                    Et[jp,0] = ackleyf(plot_points[jp,:]-x_true)
-                plt.plot(plot_points,gp_E,'b-',plot_points,Et,'g-',\
-                         mol_pos,E,'r.')
-                plt.show()
+            if c_mes>b_mes:
+                new_points = np.copy(cnew_points)
+                b_mes = c_mes
             #end if
-        #end if
-    #end if
+        #end for
+    #end for
+    
+    return new_points
+#end def find_new_points
+    
+###############################################################################
 
-    return hyper_cube_F,E_F,P_F,E0_F,P0_F
-#end def gp_sgd
+def multiscale_E(E,dE,scale_thres):
+    """
+    Multiscale normalization of E
+    
+    
+    Parameters
+    ----------
+    E:              Energy
+    dE:             Uncertainty in Energy
+    scale_thres:    Scale parameter
+    
+    Returns
+    -------
+    E_a:        Scale parameters
+    E_b:        Scale parameters
+    E_scaled:   Scaled energies
+    dE_scaled:  Scaled uncertainties in energies
+    """
+    
+    E_a = min(E)
+    E_scaled = (E-E_a)/scale_thres
+    dE_scaled = dE/scale_thres
+    us_ind = np.argwhere(E_scaled[:,0]>1)
+    if us_ind.shape[0]>0:   
+        E_scaled[us_ind[:,0],0] = np.log(E_scaled[us_ind[:,0],0])+1
+        dE_scaled[us_ind[:,0],0] = max(dE_scaled)
+        E_b = max(E_scaled[us_ind[:,0],0])
+        if E_b>2*scale_thres:
+            E_scaled[us_ind[:,0],0] = scale_thres*(E_scaled[us_ind[:,0],0]-1)/(E_b-1)+1
+        else:
+            E_b= 0
+        #end if
+    else:
+        E_scaled = E
+        dE_scaled = dE
+        E_a = 0
+        E_b= 0
+    #end if
+        
+    return E_a,E_b,E_scaled,dE_scaled
+#end def multiscale_E
+
+###############################################################################
+
+def inv_multiscale_E(E_a,E_b,E_scaled,scale_thres):
+    """
+    Reverses Multiscale normalization of E
+    
+    
+    Parameters
+    ----------
+    E_a:            Scale parameters
+    E_b:            Scale parameters
+    E_scaled:       Scaled energies
+    scale_thres:    Scale parameter
+    
+    Returns
+    -------
+    E:          Energy
+    """
+    if E_a == 0:
+        E = E_scaled
+    else:
+        us_ind = np.argwhere(E_scaled[:,0]>1)
+        if us_ind.shape[0]>0:
+            if E_b!=0:
+                E_scaled[us_ind[:,0],0] = (E_scaled[us_ind[:,0],0]-1)*(E_b-1)/scale_thres+1
+            #end if
+            E_scaled[us_ind[:,0],0] = np.exp(E_scaled[us_ind[:,0],0]-1)
+        #end if
+        E = scale_thres*E_scaled+E_a
+    #end if
+        
+    return E
+#end def inv_multiscale_E
+
+###############################################################################
+    
+def norm_pos(mol_pos,hyper_cube):
+    """
+    Normalizes molecular positions to unit hyper cube
+    
+    
+    Parameters
+    ----------
+    hyper_cube:     Hyper cube
+    mol_pos_norm:   Normalized GP centers
+    
+    
+    Returns
+    -------
+    mol_pos:        GP centers
+    """
+    
+    [n,d] = mol_pos.shape
+    mol_pos_norm = (mol_pos-np.repeat(np.reshape(hyper_cube[:,0],[1,d]),n,axis=0))/ \
+        np.repeat(np.reshape(hyper_cube[:,1]-hyper_cube[:,0],[1,d]),n,axis=0)
+
+        
+    return mol_pos_norm
+#end def norm_pos
+
+###############################################################################
+    
+def inv_norm_pos(mol_pos_norm,hyper_cube):
+    """
+    Reverses normalization of molecular positions to unit hyper cube
+    
+    
+    Parameters
+    ----------
+    hyper_cube:     Hyper cube
+    mol_pos_norm:   Normalized GP centers
+    
+    
+    Returns
+    -------
+    mol_pos:        GP centers
+    """
+    
+    [n,d] = mol_pos_norm.shape
+    mol_pos = np.repeat(np.reshape(hyper_cube[:,1]-hyper_cube[:,0],[1,d]) \
+                ,n,axis=0)*mol_pos_norm+ \
+                np.repeat(np.reshape(hyper_cube[:,0],[1,d]),n,axis=0)
+
+    return mol_pos
+#end def inv_norm_pos
 
 ###############################################################################
 
 
-
-
-
-def ackley_example():
+def gp_sdg_setup(GP_info):
     """
-    Example using Gaussian Process optimization on Ackley function
-
+    Setup parameters and variables for GP
+    
+    
+    Parameters
+    ----------
+    GP_info:    GP information 
+    E_b:        Scale parameters
+    E_scaled:   Scaled energies
+    
+    
+    Returns
+    -------
+    E:   
     """
+    
+    ## Define Empty arrays for energies, positions, and uncertainties
+    E = []
+    dE = []
+    mol_pos = []
 
-    ## Parameters ##
-    d = 2                               # Dimension of problem
-    nits = 3                            # Number of iterations
-    sigma = 1e-6                        # Uncertainty in data
-    hyper_cube = np.ones([d,2])         # Upper and lower bound for each dimension
-    hyper_cube[:,0] = -hyper_cube[:,0]
+    best_pos =[]
+    best_E = []
+    
+    if 'hyper_cube' not in GP_info:
+        error('Please Define GP_info.hyper_cube')
+    #end if
 
-    ## Set up gp_sgd structure ##
-    hyper_cube_F =[]
-    hyper_cube_F.append(hyper_cube)
-    E_F = []
-    P_F = []
-    E0_F = []
-    P0_F = []
+    d = GP_info.hyper_cube.shape[0]
 
-    ## Begin heirarchical optimization ##
-    for jits in range(nits+1):
-        if jits < nits:
-            hyper_cube_F,E_F,P_F,E0_F,P0_F = gp_sgd(hyper_cube_F,E_F,P_F,E0_F,P0_F,sigma)
+    GP_info.set_optional(
+        n_its           = 5,          # Number of maximum iterations
+        jits            = 0,          # Current iteration index
+        do_vis          = 0,          # Visualize Error
+        do_scale_energy = 0,          # Scale energy
+        scale_thres     = 0.,         # Energy scale parameter
+        Nc0             = (d+2)*(d+1),# Number of points to add per iteration
+        delta_r         = 1.0,        # Initial Radius of Convergence
+        delta_r_s       = 2.0,        # Radius shrink factor
+        n_min_d         = 5,          # Number of times to resample LHC
+        )
 
-            # Determine Energy for new set of points
-            npoints = P_F[jits].shape[0]
-            E = np.zeros([npoints,1])  
-            if jits == 0:
-                ## Randomly pick center for Ackley function
-                x_true = 0.05*(2.0*np.random.rand(1,d)-1.0)
-            #end if
+    return E,dE,mol_pos,best_pos,best_E,GP_info
+#end def gp_sdg_setup
 
-            ## Determine Energies on current points using Ackley function
-            for jp in range(npoints):
-                E[jp,0] = ackleyf(P_F[jits][jp,:]-x_true)
-            #end for
-            E_F.append(E)
+###############################################################################
+
+def gp_opt(E,dE,mol_pos,GP_info):
+    """
+    Setup parameters and variables for GP
+    
+    
+    Parameters
+    ----------
+    best_pos:   Estimated local minimum
+    best_E:     Estimated local energies
+    mol_pos:    GP centers
+    GP_info:    GP information 
+    
+    E:          Energy for GP centers
+    dE:         Uncertainty in energy for GP centers
+    mol_pos:    GP coefficents
+    GP_info:    GP information 
+    
+    Returns
+    -------
+    new_points:     New points to sample
+    best_pos:       Estimated local minimum
+    best_E:         Estimated local energies
+    GP_info:        GP information    
+    """
+    
+    ## Define Empty arrays for energies, positions, and uncertainties
+    d = GP_info.hyper_cube.shape[0]
+    
+    if GP_info.jits == 0:
+        mol_pos_norm = lhs(d,GP_info.Nc0,GP_info.n_min_d)
+        new_points = inv_norm_pos(mol_pos_norm,GP_info.hyper_cube)
+        best_pos = []
+        best_E =[]
+    else:
+        GP_info.delta_r = GP_info.delta_r/GP_info.delta_r_s
+        mol_pos_norm = norm_pos(mol_pos,GP_info.hyper_cube)
+        
+        if GP_info.do_scale_energy == 1:
+            E_a,E_b,E_scaled,dE_scaled = multiscale_E(E,dE,GP_info.scale_thres)
+            ## Best Approximation of Optimization Surface ##
+            ep,sigma,Co_GP = find_hyperparameters(mol_pos_norm,E_scaled,dE_scaled)
+            ## Find Global Min ##
+            best_pos_norm,best_E_scaled,neig_data = find_min(Co_GP,mol_pos_norm,E_scaled,dE_scaled,ep)
+            best_E = inv_multiscale_E(E_a,E_b,best_E_scaled,GP_info.scale_thres)
+            
         else:
-            # Final iteration where only minimum is found and no new points generated
-            hyper_cube_F,E_F,P_F,E0_F,P0_F = gp_sgd(hyper_cube_F,E_F,P_F,E0_F,P0_F,sigma,gen_new=0)
+            ep,sigma,Co_GP = find_hyperparameters(mol_pos_norm,E,dE)
+            best_pos_norm,best_E,neig_data = find_min(Co_GP,mol_pos_norm,E,dE,ep)
+        #end if
+
+        best_pos = inv_norm_pos(best_pos_norm,GP_info.hyper_cube)
+        
+        new_points = find_new_points(mol_pos_norm,best_pos_norm,GP_info.delta_r,GP_info.Nc0)
+        
+        new_points = inv_norm_pos(new_points,GP_info.hyper_cube)
+        
+        if GP_info.do_vis == 1 and available(plt):
+            neig_bool = neig_data[0:best_E.shape[0],:]- \
+                2*neig_data[best_E.shape[0]:2*best_E.shape[0],:]
+            neig_bool = neig_bool<np.repeat(best_E,d+1,axis=1)
+            neig_percent = np.zeros(neig_bool.shape)
+            for jx in range(neig_bool.shape[0]):
+                for jy in range(neig_bool.shape[1]):
+                    neig_percent[jx,jy] = float(neig_bool[jx,jy])
+                #end for
+            #end for
+            neig_percent = 100*np.mean(neig_percent,axis=1)
+            x = np.zeros((d+1)*best_E.shape[0])
+            y = np.zeros((d+1)*best_E.shape[0])
+            dy = np.zeros((d+1)*best_E.shape[0])
+            Ex = np.zeros(best_E.shape[0])
+            for jx in range(best_E.shape[0]):
+                x[jx*(d+1):(jx+1)*(d+1)] = jx
+                y[jx*(d+1):(jx+1)*(d+1)] = neig_data[jx,:]
+                dy[jx*(d+1):(jx+1)*(d+1)]= neig_data[jx+best_E.shape[0],:]
+                Ex[jx] = jx
+            #end for
+
+            plt.figure()
+            plt.errorbar(x, y, yerr=dy, fmt='o',label='sampled point with uncertainty')
+            plt.plot(Ex,best_E[:,0], 'r+',label='estimated minumum')
+            plt.legend()
+            plt.title('Neigborhood Information')
+            plt.xlabel('Local Minimum')
+            plt.ylabel('E')
+            plt.show()
+        elif GP_info.do_vis == 1:
+            warn('visualization requested but matplotlib is not available\nplotting suppressed')
+        #end if
+    #end if
+        
+    return new_points,best_pos,best_E,GP_info
+#end def gp_opt
+
+###############################################################################
+    
+
+def gp_opt_example():
+    """
+    Example using Gaussian Process optimization for noisy data and multiple mins
+
+    """
+
+    ## Required to define is hypercube  ##
+    d = 4                                       # Dimension of problem
+    GP_info = GParmFlags()
+    GP_info.hyper_cube = np.ones([d,2])/2       # Upper and lower bound for each dimension
+    GP_info.hyper_cube[:,0] = -GP_info.hyper_cube[:,0]/2
+
+
+    ## Parameters Used for Optimization Function Defining Range of Errors in Energy##
+    sigma_a = 1e-3;
+    sigma_b = 1e-1;
+
+
+    ## Setup Gausian Process Structure ##
+    E,dE,mol_pos,best_pos,best_E,GP_info = gp_sdg_setup(GP_info)
+
+    ## Do Gaussian Process Iteration ##
+    for jits in range(GP_info.n_its):
+        GP_info.jits = jits
+        new_points,best_pos,best_E,GP_info = gp_opt(E,dE,mol_pos,GP_info)
+        print jits,'best_P',best_pos
+        print jits,'best_E',best_E
+        if jits < GP_info.n_its-1:
+            ## Simulate random uncertainty in optimization function
+            new_dE = (sigma_b-sigma_a)*np.random.rand(GP_info.Nc0,1)+sigma_a
+            new_E = opt_surface_f(new_points)+new_dE*np.random.randn(GP_info.Nc0, 1)
+            if jits == 0:
+                mol_pos = np.copy(new_points)
+                E = np.copy(new_E)
+                dE = np.copy(new_dE)
+            else:
+                mol_pos = np.append(mol_pos,new_points,axis=0)
+                E = np.append(E,new_E,axis=0)
+                dE = np.append(dE,new_dE,axis=0)
+            #end if
         #end if
     #end for
-#end def ackley_example
+#end def gp_opt_example
 
+
+
+if __name__=='__main__':
+    gp_opt_example()
+#end if
